requirements.txt
# Runtime dependencies (flexible pins suitable for OSS consumers)

# Google GenAI SDK (Client-centric API)
google-genai~=1.28